requirements.txt
streamlit
openai
pandas
numpy
python-dotenv
scikit-learn
sentence-transformers